Refactor: implement PrivateRoute to maintain login
