+++ FE/src/hooks/useMaintainLogin.tsx
@@ -28,7 +28,6 @@ const useMaintainLogin = () => {
       setUserState({ name, email, profileImageUrl });
       setCookie(REFRESH_TOKEN, refreshToken, REFRESH_TOKEN_OPTIONS);
       axios.defaults.headers.common.Authorization = `Bearer ${accessToken}`;
-      navigate('/');
     } catch (error) {
       console.log(error);
       navigate('/login');
