+++ FE/src/pages/Home/index.tsx
@@ -1,26 +1,15 @@
-import { useEffect } from 'react';
 import { useRecoilValue } from 'recoil';
 
 import Loading from '@/components/common/Loading';
 import HomeHeader from '@/components/HomeHeader';
 import IssueList from '@/components/IssueList';
-import { useIsLoggedIn } from '@/hooks/useIsLoggedIn';
 import { useGetIssue } from '@/hooks/useIssue';
-import useMovePage from '@/hooks/useMovePage';
 import { issueStatusState } from '@/stores/atoms/issue';
 import { InnerContainer, MainWrapper } from '@/styles/common';
 
 const Home = () => {
   const issueStatus = useRecoilValue(issueStatusState);
   const { data, isLoading } = useGetIssue(issueStatus);
-  const isLoggedIn = useIsLoggedIn();
-  const [goLogin] = useMovePage('/login');
-
-  useEffect(() => {
-    if (!isLoggedIn) {
-      goLogin();
-    }
-  }, [useIsLoggedIn]);
 
   return (
     <MainWrapper>
